move actuators to new subdirectory
