+++ pkg/actuators/azure/actuator_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/golang/mock/gomock"
 	openshiftapiv1 "github.com/openshift/api/config/v1"
+	"github.com/openshift/cloud-credential-operator/pkg/actuators/azure"
+	azuremock "github.com/openshift/cloud-credential-operator/pkg/actuators/azure/mock"
 	minterv1 "github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1"
-	"github.com/openshift/cloud-credential-operator/pkg/azure"
-	azuremock "github.com/openshift/cloud-credential-operator/pkg/azure/mock"
 	annotatorconst "github.com/openshift/cloud-credential-operator/pkg/operator/secretannotator/constants"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
