+++ pkg/actuators/azure/mock/client_generated.go
