+++ pkg/operator/credentialsrequest/credentialsrequest_controller_gcp_test.go
@@ -37,11 +37,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	mintergcp "github.com/openshift/cloud-credential-operator/pkg/actuators/gcp"
+	"github.com/openshift/cloud-credential-operator/pkg/actuators/gcp/actuator"
+	mockgcp "github.com/openshift/cloud-credential-operator/pkg/actuators/gcp/mock"
 	"github.com/openshift/cloud-credential-operator/pkg/apis"
 	minterv1 "github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1"
-	mintergcp "github.com/openshift/cloud-credential-operator/pkg/gcp"
-	"github.com/openshift/cloud-credential-operator/pkg/gcp/actuator"
-	mockgcp "github.com/openshift/cloud-credential-operator/pkg/gcp/mock"
 	annotatorconst "github.com/openshift/cloud-credential-operator/pkg/operator/secretannotator/constants"
 	gcpconst "github.com/openshift/cloud-credential-operator/pkg/operator/secretannotator/gcp"
 	"github.com/openshift/cloud-credential-operator/pkg/operator/utils"
