+++ pkg/operator/credentialsrequest/credentialsrequest_controller_test.go
@@ -39,11 +39,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	minteraws "github.com/openshift/cloud-credential-operator/pkg/actuators/aws"
+	"github.com/openshift/cloud-credential-operator/pkg/actuators/aws/actuator"
+	mockaws "github.com/openshift/cloud-credential-operator/pkg/actuators/aws/mock"
 	"github.com/openshift/cloud-credential-operator/pkg/apis"
 	minterv1 "github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1"
-	minteraws "github.com/openshift/cloud-credential-operator/pkg/aws"
-	"github.com/openshift/cloud-credential-operator/pkg/aws/actuator"
-	mockaws "github.com/openshift/cloud-credential-operator/pkg/aws/mock"
 	"github.com/openshift/cloud-credential-operator/pkg/operator/credentialsrequest/constants"
 	annotatorconst "github.com/openshift/cloud-credential-operator/pkg/operator/secretannotator/constants"
 	"github.com/openshift/cloud-credential-operator/pkg/operator/utils"
